Fixed issue #42
