+++ Controller/Annotations/NamePrefix.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace FOS\RestBundle\Controller\Annotations;
+
+use Doctrine\Common\Annotations\Annotation;
+
+/*
+ * This file is part of the FOSRestBundle
+ *
+ * (c) Lukas Kahwe Smith <smith@pooteeweet.org>
+ * (c) Konstantin Kudryashov <ever.zet@gmail.com>
+ * (c) Bulat Shakirzyanov <mallluhuct@gmail.com>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+/**
+ * NamePrefix Route annotation class.
+ */
+class NamePrefix extends Annotation
+{
+}
