+++ Controller/Annotations/Put.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace FOS\RestBundle\Controller\Annotations;
+
+/*
+ * This file is part of the FOSRestBundle
+ *
+ * (c) Lukas Kahwe Smith <smith@pooteeweet.org>
+ * (c) Konstantin Kudryashov <ever.zet@gmail.com>
+ * (c) Bulat Shakirzyanov <mallluhuct@gmail.com>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+/**
+ * PUT Route annotation class.
+ */
+class Put extends Route
+{
+    public function getMethod()
+    {
+        return 'PUT';
+    }
+}
