+++ Resources/config/routing.xml
@@ -9,7 +9,6 @@
         <parameter key="fos_rest.routing.loader.controller.class">FOS\RestBundle\Routing\Loader\RestRouteLoader</parameter>
         <parameter key="fos_rest.routing.loader.yaml_collection.class">FOS\RestBundle\Routing\Loader\RestYamlCollectionLoader</parameter>
         <parameter key="fos_rest.routing.loader.xml_collection.class">FOS\RestBundle\Routing\Loader\RestXmlCollectionLoader</parameter>
-        <parameter key="fos_rest.routing.loader.controller.annotations_namespace">FOS\RestBundle\Controller\Annotations\</parameter>
 
     </parameters>
 
@@ -18,14 +17,7 @@
         <service id="fos_rest.routing.loader.controller" class="%fos_rest.routing.loader.controller.class%">
             <argument type="service" id="service_container" />
             <argument type="service" id="controller_name_converter" />
-            <argument type="service">
-                <service class="Doctrine\Common\Annotations\AnnotationReader">
-                    <call method="setAnnotationNamespaceAlias">
-                        <argument>%fos_rest.routing.loader.controller.annotations_namespace%</argument>
-                        <argument>rest</argument>
-                    </call>
-                </service>
-            </argument>
+            <argument type="service" id="annotation_reader" />
             <tag name="routing.loader" />
         </service>
